Constellation Works
I added the Cancer constellation (example). If you guys wanna add a real description of it, along with an image and maybe if you wanna make it prettier? Also, i forgot to make the backpack able to open with the press of a button. you can do that in the GameController script.